Move the live-inert projector into the workflow package
The workflow child must project a definition to its inert form to
re-verify it on load, but the child's package must not depend on the
deploy package where that projector lived. Move it into the workflow
package, which the child already depends on, so the projection is
reachable without crossing that boundary.
